% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepNetwork.R
\name{DeepNetworkPredict}
\alias{DeepNetworkPredict}
\title{Predictions from a simple deep network.}
\usage{
DeepNetworkPredict(x, model, verbose = FALSE)
}
\arguments{
\item{x}{matrix with input data}

\item{model}{a model trained with the deep network trainer.}

\item{verbose}{verbose output?}
}
\description{
This will do prediction using a trained simple 'deep' neural network with two hidden layers.
For more information refer to the Shark tutorial at http://image.diku.dk/shark/sphinx_pages/build/html/rest_sources/tutorials/algorithms/deep_denoising_autoencoder_network.html
}
