% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppShark.package.skeleton.R
\name{RcppShark.package.skeleton}
\alias{RcppShark.package.skeleton}
\title{create a skeleton for a package using RcppShark}
\usage{
RcppShark.package.skeleton(name = "anRpackage", list = character(),
  environment = .GlobalEnv, path = ".", force = FALSE,
  code_files = character(), example_code = TRUE, author = "Your Name",
  maintainer = if (missing(author)) "Your Name" else author,
  email = "your@email.com", license = "GPL (>= 2)")
}
\arguments{
\item{name}{character string: the package name and directory name for your package.}

\item{list}{character vector naming the R objects to put in the package. 
Usually, at most one of "list", "environment", or "code_files" will be supplied.  See "Details" of package.skeleton.}

\item{environment}{an environment where objects are looked for.  See "Details" of package.skeleton.}

\item{path}{path to put the package directory in.}

\item{force}{If "FALSE" will not overwrite an existing directory.}

\item{code_files}{a character vector with the paths to R code files to build the package around.  See "Details" of package.skeleton.}

\item{example_code}{add example code to package?}

\item{author}{author in the created DESCRIPTIONS file}

\item{maintainer}{maintainer in the created DESCRIPTIONS file}

\item{email}{email in the created DESCRIPTIONS file}

\item{license}{licence in the created DESCRIPTIONS file}
}
\description{
create a skeleton for a package using RcppShark
}
