% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppShark.R
\docType{package}
\name{RcppShark}
\alias{RcppShark}
\alias{RcppShark-package}
\title{RcppShark: An R interface to the C++/Boost Shark machine learning library.}
\description{
The RcppShark package is a wrapper for the Shark machine learning library. 
Shark is programmed in C++/Boost and contains hundred of classes and methods.
}
\section{RcppShark functions}{

RcppShark.package.skeleton
DeepNetworkTrain
DeepNetworkPredict
SharkBSGDTrain
SharkBSGDPredict
}

